Editorial pass up to and including "Requirements Language" Section (up to line 271). Some minor corrections. Two comments
- (Lines 183-204) [content]:
We list the different modes: client-server, symmetric, control and broadcast. We then proceed to talk about all those that used to be in the last iteration, but not broadcast. For consistence, I think we should either eliminate discussion of symmetric and control modes, or insert discussion of broadcast mode.

- (Lines 213 and 224) [editorial]:
Inconsistent quotation marks, should be made consistent. (I would have done it, but just don't know which is better/more suitable)