Simplify the NTPv4 Server Negotiation record
RFC3986 allows the use of percent encoding in host names which
would require a client to parse the percent encoding, convert
the percent encoding to Unicode and finally convert the unicode
code points to an A-LABEL which can be looked up with DNS.

Remove the reference to RFC3986 and specify the possible contents
explicitly.  The new specification does not allow percent encoding
which makes the client simpler by specifying that if
internationalized labels are needed in the host name they must
be A-LABEL encoded already.